Move role specific `Tunnel` impls to dedicated modules
